% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{ConvertNDSIVectorToMatrix}
\alias{ConvertNDSIVectorToMatrix}
\title{Convert from NDSI vector to matrix}
\usage{
ConvertNDSIVectorToMatrix(x)
}
\arguments{
\item{x}{The measurement vector. Labels are freqA_freqB format}
}
\value{
A matrix with freqA on rows, freqB on cols, with NA
}
\description{
Convert from NDSI vector to matrix
}

