% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{ConvertSpectrumToNDSI}
\alias{ConvertSpectrumToNDSI}
\title{Convert from spectrum to NDSI}
\usage{
ConvertSpectrumToNDSI(spectrum)
}
\arguments{
\item{spectrum}{An input spectrum. First column is species names.}
}
\value{
A data.frame with species names (first column) and NDSI columns
}
\description{
Convert from spectrum to NDSI
}
\details{
Input columns have colnames that have wavelength in nm (as character), other than species names
Export columns have colnames that have the pair: 700_805 for NDSI for 700 nm and 805 nm bands
}

