% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{DoUnivariateOUwieDredge}
\alias{DoUnivariateOUwieDredge}
\title{Fit various BM1 and OU1 to a single trait, using measurement error}
\usage{
DoUnivariateOUwieDredge(data, phy)
}
\arguments{
\item{data}{A data.frame with columns for species names, species mean, and species standard error}

\item{phy}{A phylo object.}

\item{reconcile}{If TRUE, use geiger's treedata() function to prune phylogeny and data to same leaf set}
}
\value{
A data.frame with returned parameters
}
\description{
Fit various BM1 and OU1 to a single trait, using measurement error
}
\details{
The returned object includes the likelihoods, AIC, and AIC weights of BM1 and OU1, sigma for each,
alpha and theta for OU, and model-averaged estimates for sigma and alpha.
}

