% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{GetPhylogeneticMeans}
\alias{GetPhylogeneticMeans}
\title{Get phylogenetic means for species (including ones with missing data)}
\usage{
GetPhylogeneticMeans(data, phy, model = "BM", phylo_correlated = FALSE,
  pad.missing = TRUE, ...)
}
\arguments{
\item{data}{A data.frame with first column species name, other columms trait data. NA are allowed}

\item{phy}{A tree for all species in the data.}

\item{model}{A model for phylopars. Options are "BM", "OU" "lambda", "kappa", "delta", "EB", "star"}

\item{phylo_correlated}{If TRUE, parameters are estimated assuming traits are correlated. This can be quite slow.}

\item{...}{Other options to pass to phylopars()}
}
\value{
list containing means and variances for species, as well as the phylopars object
}
\description{
Get phylogenetic means for species (including ones with missing data)
}
\details{
Note that this will use same model (but different params) for each column.
If you want to use different models, do different subsets of the data
}

