% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{GetSpeciesMeansWithError}
\alias{GetSpeciesMeansWithError}
\title{Aggregate by species. Return mean and standard error}
\usage{
GetSpeciesMeansWithError(data)
}
\arguments{
\item{data}{A data.frame: first column species, second column trait data}
}
\value{
A data.frame with species name, species mean, and species standard error
}
\description{
Aggregate by species. Return mean and standard error
}

