% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{PadData}
\alias{PadData}
\title{Add NA for missing species}
\usage{
PadData(data, taxa, strict = FALSE)
}
\arguments{
\item{data}{A data.frame: first column species, remaining columns trait data}

\item{taxa}{A list of species that should be in the final data.frame. Missing ones will be added}

\item{strict}{If TRUE, only species listed in taxa will be in output. Otherwise, species will be added but not deleted}
}
\value{
A data.frame: first column species, remaining columns trait data, with NA for some taxa
}
\description{
Add NA for missing species
}

