% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotNDSI}
\alias{PlotNDSI}
\title{Plot parameter with NDSI}
\usage{
PlotNDSI(x)
}
\arguments{
\item{x}{The vector of measurements. Labels are freqA_freqB format.}
}
\value{
Nothing, though it creates a heatmap Plot
}
\description{
Plot parameter with NDSI
}

