% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{ConvertSpectrumToNDSI}
\alias{ConvertSpectrumToNDSI}
\title{Convert from spectrum to NDSI}
\usage{
ConvertSpectrumToNDSI(spectrum)
}
\arguments{
\item{spectrum}{An input spectrum. First column is species names.}
}
\value{
A data.frame with species names (first column) and NDSI columns
}
\description{
Convert from spectrum to NDSI
}

